requirements.txt
asgiref==3.2.10
bootstrap-admin==0.4.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.1
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
idna==2.10
Pillow==8.0.1
pipreqs==0.4.10
psycopg2==2.8.6
pycparser==2.20
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.25.11
whitenoise==5.2.0
yarg==0.1.9
